+++ src/TaskRunner/Commands/DrupalCommands.php
@@ -145,12 +145,12 @@ class DrupalCommands extends Drupal8Commands
   'prefix' => getenv('DRUPAL_DATABASE_PREFIX'),
   'host' => getenv('DRUPAL_DATABASE_HOST'),
   'port' => getenv('DRUPAL_DATABASE_PORT'),
-  'namespace' => 'Drupal\\\\Core\\\\Database\\\\Driver\\\\mysql',
-  'driver' => 'mysql',
+  'namespace' => getenv('DRUPAL_DATABASE_DRIVER') !== FALSE ? 'Drupal\\\\Core\\\\Database\\\\Driver\\\\' . getenv('DRUPAL_DATABASE_DRIVER') : 'Drupal\\\\Core\\\\Database\\\\Driver\\\\mysql',
+  'driver' => getenv('DRUPAL_DATABASE_DRIVER') !== FALSE ? getenv('DRUPAL_DATABASE_DRIVER') : 'mysql',
 );
 
 // Location of the site configuration files, relative to the site root.
-\$config_directories['sync'] = '../config/sync';
+\$settings['config_sync_directory'] = '../config/sync';
 
 \$settings['hash_salt'] = getenv('DRUPAL_HASH_SALT') !== FALSE ? getenv('DRUPAL_HASH_SALT') : '{$hashSalt}';
 \$settings['file_private_path'] =  getenv('DRUPAL_PRIVATE_FILE_SYSTEM') !== FALSE ? getenv('DRUPAL_PRIVATE_FILE_SYSTEM') : 'sites/default/private_files';