+++ src/TaskRunner/Commands/InstallCommands.php
@@ -14,6 +14,7 @@ use Symfony\Component\Console\Input\InputOption;
 class InstallCommands extends AbstractCommands
 {
 
+    use TaskRunnerTasks\CollectionFactory\loadTasks;
     use TaskRunnerTasks\Drush\loadTasks;
 
     /**
@@ -70,6 +71,12 @@ class InstallCommands extends AbstractCommands
         $runner_bin = $this->getConfig()->get('runner.bin_dir') . '/run';
         $tasks[] = $this->taskExec($runner_bin . ' toolkit:install-dump');
         $tasks[] = $this->taskExec($runner_bin . ' toolkit:run-deploy');
+
+        // Collect and execute list of commands set on local runner.yml.
+        $commands = $this->getConfig()->get("toolkit.install.clone.commands");
+        if (!empty($commands)) {
+            $tasks[] = $this->taskCollectionFactory($commands);
+        }
 
         // Build and return task collection.
         return $this->collectionBuilder()->addTaskList($tasks);